+++ packages/transport-grpc/CHANGELOG.md
@@ -1,4 +1,19 @@
-### Unreleased
+## 1.0.0-alpha.1
+
+### Major Changes
+
+- Release 1.0.0 alpha
+
+### Patch Changes
+
+- Updated dependencies [7469c5c3]
+- Updated dependencies
+
+  - @onflow/util-address@1.0.0-alpha.0
+  - @onflow/util-invariant@1.0.0-alpha.0
+  - @onflow/protobuf@1.0.0-alpha.0
+  - @onflow/rlp@1.0.0-alpha.0
+  - @onflow/util-template@1.0.0-alpha.0
 
 - 2022-03-16 -- [@bthaile](https://github.com/bthaile) Payer can now be an array.
 - 2022-02-11 -- Uses Buffer as provided by context injection